refactor ui to page view
