+++ infra/foundation-web/template.yaml
@@ -41,6 +41,21 @@ Conditions:
     - !Not [!Equals [!Ref DomainName, '']]
     - !Not [!Equals [!Ref DomainHostedZoneId, '']]
 
+Metadata:
+  esbuild-properties: &esbuild-properties
+    Format: esm
+    Minify: true
+    OutExtension:
+      - .js=.mjs
+    Target: es2020
+    Sourcemap: false
+    EntryPoints:
+      - index.mjs
+    Banner:
+      - js=import { createRequire } from 'module'; const require = createRequire(import.meta.url);
+    External:
+      - '@aws-sdk/*'
+
 Resources:
   FoundationUserPool:
     Type: AWS::Cognito::UserPool
@@ -84,18 +99,9 @@ Resources:
     Metadata:
       BuildMethod: esbuild
       BuildProperties:
-        Format: esm
-        Minify: true
-        OutExtension:
-          - .js=.mjs
-        Target: es2020
-        Sourcemap: false
+        <<: *esbuild-properties
         EntryPoints:
           - post-confirmation.mjs
-        Banner:
-          - js=import { createRequire } from 'module'; const require = createRequire(import.meta.url);
-        External:
-          - '@aws-sdk/*'
     Properties:
       CodeUri: functions
       Handler: post-confirmation.handler
